EES-7273 - tidy up post PR feedback
