+++ tests/robot-tests/tests/libs/snapshot_notification.py
@@ -9,12 +9,6 @@ channels receive the same message.
 from dataclasses import dataclass
 from typing import Optional
 
-from slack_sdk.webhook import WebhookClient
-from tests.libs.logger import get_logger
-from tests.libs.teams import TeamsService
-
-logger = get_logger(__name__)
-
 PULL_REQUESTS_URL = "https://github.com/dfe-analytical-services/explore-education-statistics/pulls/dfe-sdt"
 
 # Adaptive Card colours. Slack headings cannot be coloured, so these apply to Teams only.
@@ -23,7 +17,7 @@ ATTENTION = "attention"
 
 
 @dataclass(frozen=True)
-class Notification:
+class SnapshotNotification:
     title: str
     body: str
     colour: str = "default"
@@ -80,8 +74,8 @@ class Notification:
         }
 
 
-def snapshots_do_not_match() -> Notification:
-    return Notification(
+def snapshots_do_not_match() -> SnapshotNotification:
+    return SnapshotNotification(
         title="⚠️ Snapshots do not match",
         body="See {link} for more details",
         colour=WARNING,
@@ -90,47 +84,9 @@ def snapshots_do_not_match() -> Notification:
     )
 
 
-def snapshot_process_failed(message: str) -> Notification:
-    return Notification(
+def snapshot_process_failed(message: str) -> SnapshotNotification:
+    return SnapshotNotification(
         title="❌ Snapshot process failed",
         body=str(message),
         colour=ATTENTION,
     )
-
-
-class SnapshotNotifier:
-    """
-    Sends a notification to every channel it has been given a webhook URL for.
-    Channels without a webhook URL are skipped, so that the script can be run
-    locally without notifying anyone.
-    """
-
-    def __init__(self, slack_webhook_url: Optional[str] = None, teams_webhook_url: Optional[str] = None):
-        self.slack_webhook_url = slack_webhook_url
-        self.teams_service = TeamsService(teams_webhook_url) if teams_webhook_url else None
-
-    def send(self, notification: Notification) -> None:
-        # Teams is notified first as it never raises, so a Slack outage cannot
-        # stop the message reaching Teams.
-        self._send_to_teams(notification)
-        self._send_to_slack(notification)
-
-    def _send_to_teams(self, notification: Notification) -> None:
-        if not self.teams_service:
-            logger.info("No Teams webhook URL was given; skipping Teams notification")
-            return
-
-        self.teams_service.send_snapshot_notification(notification.to_teams_card())
-
-    def _send_to_slack(self, notification: Notification) -> None:
-        if not self.slack_webhook_url:
-            logger.info("No Slack webhook URL was given; skipping Slack notification")
-            return
-
-        response = WebhookClient(self.slack_webhook_url).send(
-            text=notification.title, blocks=notification.to_slack_blocks()
-        )
-
-        assert (
-            response.status_code == 200 and response.body == "ok"
-        ), f"Slack notification failed with status code: {response.status_code}"
